% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmodel.R
\name{FR.model}
\alias{FR.model}
\title{Factorial Regression (FR) models with environmental or genotypic covariates}
\usage{
FR.model(df.y, intercept = TRUE)
}
\arguments{
\item{df.y}{is a data.frame containing the following colunms: environment (factor), genotype (factor), GxE or G+GxE effects (numeric) and covariates (numeric).}

\item{intercept}{TRUE (default) or FALSE if considering include fixed genotypic intercept}
}
\description{
Returns factorial regression from phenotypic and environmental data.
The number of phenotypic observations can not exceed the number of parameters
(coefficients of the genotypes for each covariate) to be estimated.
Note: the model is based on ordinary least squares (OLS) and multicollinearity effects
are not removed from the analysis.
Future updates of the function will allow FR modeling based on Bayesian inference (Gibbs-sampling) and partial least squares (PLS).
}
\references{
Costa Neto GMF. Integrating environmental covariates and thematic maps into genotype by environment interaction analysis in upland rice. Master degree Thesis in Genetics and Plant Breeding, Agronomy School, Federal University of Goiás. Brazil, 2017. 122f.

Van Eeuwijk FA. Linear and bilinear models for the analysis of multi-environment trials: I. An inventory of models. Euphytica. 1995;84(1):1–7.

Brancourt-Hulmel M, Denis JB, Lecomte C. Determining environmental covariates which explain genotype environment interaction in winter wheat through probe genotypes and biadditive factorial regression. Theor Appl Genet. 2000;100(2):285–98.

Balfourier F, Oliveira JA, Charmet G, Arbones E. Factorial regression analysis of genotype by environment interaction in ryegrass populations, using both isozyme and climatic data as covariates. Euphytica. 1997;98(1):37–46.

Baril CP, Denis J-B, Wustman R, Van Eeuwijk FA. Analysing genotype by environment interaction in Dutch potato variety trials using factorial regression. Euphytica. 1995;84(1):23–9.

Vargas M, Crossa J, Van Eeuwijk FA, Ramírez ME, Sayre K. Using partial least squares regression, factorial regression, and AMMI models for interpreting genotype x environment interaction. Crop Sci. 1999;39(4):955–67.
}
\author{
Germano M F Costa Neto
}
