% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRsampling.R
\name{FR.sampling}
\alias{FR.sampling}
\title{Sampling tools for cross-validation}
\usage{
FR.sampling(df.y, f = 0.2, part.env = 1, intercept = TRUE)
}
\arguments{
\item{df.y}{is a data.frame containing the following colunms: environment (factor), genotype (factor), GxE or G+GxE effects (numeric) and covariates (numeric).}

\item{f}{is a fraction of genotypes that must be used as validation set (default f = 0.2, i.e., 20%)}

\item{part.env}{is the number of environments that be leave out of the training set (0 > part.env > number of environments). Default is 1 (leave-one-out scheme).}

\item{intercept}{TRUE (default) or FALSE if considering include fixed genotypic intercept}

\item{boot}{is the number of boots (default is equal to 1000)}
}
\description{
Sampling tools for cross-validation
}
\author{
Germano M F Costa Neto
}
