% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoRaster.R
\name{geo.rasters}
\alias{geo.rasters}
\title{Build up raster stack files of environmental covariables}
\usage{
geo.rasters(raster.file, layer.names = c("Longitude", "Latitude",
  "Elevation"), main = "Elevation", scale = T,
  crs = "+proj=longlat +ellps=WGS84 +no_defs")
}
\arguments{
\item{raster.file}{}
}
\description{
Build up raster stack files of environmental covariables
}
\author{
Germano M F Costa Neto
}
