% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdp.R
\name{plot.adp}
\alias{plot.adp}
\title{ploting yield adaptability maps from raster files}
\usage{
\method{plot}{adp}(adp.raster, nclass = 10, plot = c("full range",
  "positive", "negative"), shp = NULL, p.strip = list(cex = 0.7, lines
  = 1, col = "black"), colours = c("blue3", "royalblue", "cyan2",
  "green3", "yellow", "red"))
}
\arguments{
\item{nclass}{numeric, is a number of classes for legend values}

\item{plot}{is a type of plot, using all values (full range), only positive values (positive) or negative (negative)}

\item{shp}{is a additional shape file}

\item{dp.raster}{is a raster or a raster stack file for predicted adp values of the genotypes.}
}
\description{
ploting yield adaptability maps from raster files
}
\author{
Germano M F Costa Neto
}
