% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicAd.R
\name{predict.Ad}
\alias{predict.Ad}
\title{predicting yield adaptability from genotypic sensibility coefficients and raster files.}
\usage{
\method{predict}{Ad}(coef, cov.raster, df.x = NULL, yx = NULL,
  intercept = T)
}
\arguments{
\item{coef}{data.frame containing genotype id (gid) and genotypic sensibility coefficients}

\item{cov.raster}{is a raster or a raster stack file for environmental covariates}

\item{df.x}{data.frame containing environmental covarites per pixel (used when cov.raster is NULL)}

\item{yx}{data.frame containing geographic coordinates (latitude and longitude) per pixel (used when cov.raster is NULL)}

\item{intercept}{TRUE (default) or FALSE if considering include fixed genotypic intercept}
}
\description{
predicting yield adaptability from genotypic sensibility coefficients and raster files.
}
\author{
Germano M F Costa Neto
}
