% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFR.R
\name{predict.FR}
\alias{predict.FR}
\title{Predicting yield adaptability values for new locations}
\usage{
\method{predict}{FR}(fr.model, new.trial, intercept = TRUE)
}
\arguments{
\item{fr.model}{is a object containing FR.model output.}

\item{new.trial}{is a dataframe containing geographic covariates of new trials}

\item{intercept}{TRUE (default) or FALSE if considering include fixed genotypic intercept}
}
\description{
Predicting yield adaptability values for new locations
}
\author{
Germano M F Costa Neto
}
