% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-by-trial.R
\name{cov.by.trial}
\alias{cov.by.trial}
\title{Collects environmental covariables from rasters files and geographic information}
\usage{
cov.by.trial(cov.raster, reference, long, latd, trial)
}
\arguments{
\item{cov.raster}{is a raster stack object containing layers of environmental covariates}

\item{reference}{is a data.frame containing geographic coordinate information for each location or test to be sampled}

\item{long}{is a latitude reference vector within referebce data.frame}

\item{latd}{is a latitude reference vector within referebce data.frame}

\item{trial}{is a trial or locaion id within referebce data.frame}
}
\description{
Collects environmental covariables from rasters files and geographic information
}
\author{
Germano M F Costa Neto
}
