% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRsampling.R
\name{FR.sampling}
\alias{FR.sampling}
\title{Sampling tools for cross-validation}
\usage{
FR.sampling(df.y, f = 0.2, part.env = 1, intercept = TRUE)
}
\description{
Sampling tools for cross-validation
}
